Update PCF multi tin performer error to extend to CPC+
