+++ commons/src/main/java/gov/cms/qpp/conversion/model/error/ProblemCode.java
@@ -191,7 +191,7 @@ public enum ProblemCode implements LocalizedProblem {
 	PCF_TOO_FEW_QUALITY_MEASURE_CATEGORY(103, "PCF Submissions must have at least `(PCF Measure minimum)` "
 		+ "of the following measures: `(Listing of valid measure ids)`", true),
 	CPC_PLUS_NO_PI(104, "The file contains Promoting Interoperability (PI) data; PI data must not be reported for CPC+"),
-	PCF_MULTI_TIN_NPI_SINGLE_PERFORMER(105, "If multiple TINs/NPIs are submitted, each must be reported within a separate performer"),
+	PCF_CPC_MULTI_TIN_NPI_SINGLE_PERFORMER(105, "If multiple TINs/NPIs are submitted, each must be reported within a separate performer"),
 	PCF_NO_PI(106, "PI submissions are not allowed within PCF")
 	;
