+++ converter/src/main/java/gov/cms/qpp/conversion/validate/CpcClinicalDocumentValidator.java
@@ -1,14 +1,23 @@
 package gov.cms.qpp.conversion.validate;
 
+import org.apache.commons.lang3.StringUtils;
+import org.jdom2.Element;
+import org.jdom2.Namespace;
+import org.jdom2.filter.Filters;
+import org.jdom2.located.LocatedElement;
+import org.jdom2.xpath.XPathExpression;
+import org.jdom2.xpath.XPathFactory;
+
 import gov.cms.qpp.conversion.Context;
+import gov.cms.qpp.conversion.correlation.PathCorrelator;
 import gov.cms.qpp.conversion.decode.ClinicalDocumentDecoder;
 import gov.cms.qpp.conversion.model.Node;
 import gov.cms.qpp.conversion.model.Program;
 import gov.cms.qpp.conversion.model.TemplateId;
 import gov.cms.qpp.conversion.model.Validator;
 import gov.cms.qpp.conversion.model.error.Detail;
-import gov.cms.qpp.conversion.model.error.ProblemCode;
 import gov.cms.qpp.conversion.model.error.LocalizedProblem;
+import gov.cms.qpp.conversion.model.error.ProblemCode;
 import gov.cms.qpp.conversion.util.EnvironmentHelper;
 
 import java.time.Clock;
@@ -18,8 +27,8 @@ import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
 import java.util.List;
-
-import org.apache.commons.lang3.StringUtils;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 /**
  * Validates the Clinical Document level node for the given program: CPC+
@@ -103,6 +112,7 @@ public class CpcClinicalDocumentValidator extends NodeValidator {
 		if (hasTinAndNpi(node)) {
 			validateNumberOfTinsAndNpis(node);
 			validateApmNpiCombination(node);
+			validateSingleTinNpiPerPerformer(node);
 		}
 		validateCehrtId(node);
 	}
@@ -208,5 +218,24 @@ public class CpcClinicalDocumentValidator extends NodeValidator {
 				0, 0, 0, 0, ZoneId.of("US/Eastern"));
 		}
 		return ZonedDateTime.parse(endDate, INPUT_END_DATE_FORMAT);
+	}
+
+	private void validateSingleTinNpiPerPerformer(Node node) {
+		String performersXpath = PathCorrelator.getXpath(TemplateId.CLINICAL_DOCUMENT.toString(), "performer", Namespace.NO_NAMESPACE.getURI());
+		XPathExpression<Element> expression = XPathFactory.instance().compile(performersXpath, Filters.element(), null,  Namespace.NO_NAMESPACE);
+		List<Element> performers = expression.evaluate(node.getElementForLocation());
+		// Should be maximum of one TIN and one NPI per performer.
+		for (Element performer : performers) {
+			List<Element> tins = performer.getChildren().stream().filter(filterTinNpi("representedOrganization")).collect(Collectors.toList());
+			List<Element> npis = performer.getChildren().stream().filter(filterTinNpi("id")).collect(Collectors.toList());
+			forceCheckErrors(node).predicate(ProblemCode.PCF_CPC_MULTI_TIN_NPI_SINGLE_PERFORMER, tins.size() <= 1);
+			forceCheckErrors(node).predicate(ProblemCode.PCF_CPC_MULTI_TIN_NPI_SINGLE_PERFORMER, npis.size() <= 1);
+		}
+	}
+
+	private Predicate<Element> filterTinNpi(String name) {
+		return (Element child) -> (child.getContent().stream()
+			.filter(c -> c.getClass().equals(LocatedElement.class) && ((LocatedElement) c).getName().equals(name))
+			.toArray().length >= 1);
 	}
 }
